backend/requirements.txt
fastapi
uvicorn
spotipy
pandas
matplotlib
python-dotenv
redis